pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-formatter-agent"
version = "0.1.0"
description = "An AI agent that reformats CVs for ATS compatibility using the Strands Agents SDK."
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=1.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
cv-formatter = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]
